# knx/proxy/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "KNX Proxy API",
    "version": "1.0.0",
    "description": "REST interface for interacting with KNX group addresses via knxproxy."
  },
  "paths": {
...
    "/v1/group": {
      "post": {
        "summary": "Send a KNX group command",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Command accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupResponse"
                }
              }
            }
          },
          "4XX": {
            "description": "Invalid request"
          },
          "5XX": {
            "description": "KNX gateway error"
          }
        }
      }
    },
...
    "/v1/events": {
      "get": {
        "summary": "Retrieve recently observed events",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Maximum number of events to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Recent events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GroupRequest": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string",
            "description": "KNX group command (write, read, response)"
          },
          "destination": {
            "type": "string",
            "description": "Group address (1/2/3) or catalog name"
          },
          "source": {
            "type": "string",
            "description": "Optional source individual address"
          },
          "data": {
            "type": "string",
            "format": "byte",
            "description": "Encoded group data"
          },
          "wait_for_response": {
            "type": "boolean",
            "description": "Wait for the next response after sending"
          },
          "timeout": {
            "type": "string",
            "description": "Duration to wait for responses"
          }
        },
        "required": [
          "command",
          "destination"
        ]
      },
      "GroupResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/Event"
          }
        }
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "command": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "destination": {
            "type": "string"
          },
          "data": {
            "type": "string",
            "format": "byte"
          },
          "group": {
            "type": "string"
          },
          "dpt": {
            "type": "string"
          },
          "decoded": {
            "type": "string"
          }
        }
      }
    }
  }
}